Zipbab/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
mysql-client==0.0.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
